Hackish way to make upper_before_lower work also with CLDR tailorings
We simply change the case of each character in the $str argument of
getSortKey :) Should work in 99% of cases, but needs more thought for
special cases (e.g. if contractions have chars with diff. casing?).
The previous algorithm was modifying 3rd level weights based on the
DUCET values for lowercase vs uppercase, which no longer hold when
localized collation rules are applied.